change logic dealing cards and add RoomPokerCheckStopGameSystem

+++ server/Code/MorpehFeatures/RoomPokerFeature/Configs/RoomPokerSettingsConfig.cs
@@ -8,7 +8,7 @@ public class RoomPokerSettingsConfig
     [JsonProperty("start_game_time")] public int StartGameTime;
     [JsonProperty("player_turn_time")] public int PlayerTurnTime;
     [JsonProperty("player_turn_fast_time")] public int PlayerTurnTimeFast;
-    [JsonProperty("dealing_cards_time")] public int DealingCardsTime;
+    [JsonProperty("dealing_cards_time")] public float DealingCardsTime;
     [JsonProperty("dealing_card_time_to_table")] public int DealingCardTimeToTable;
     [JsonProperty("delay_pay_out")] public int DelayPayOut;
     [JsonProperty("delay_showdown")] public int DelayShowdown;
